+++ src/developer-platform/contracts/platform-operations-adoption-profile.js
@@ -1,0 +1,53 @@
+export class PlatformOperationsAdoptionProfile {
+  constructor({
+    platformName,
+    objectives = [], principles = [], productOwnershipFields = [], ownershipRoles = [], serviceCommitmentFields = [],
+    criticalCapabilities = [], commandControlFields = [], controlLoopStages = [], healthFields = [],
+    healthEndpointTypes = [], observabilityCorrelationFields = [], signalTypes = [], logRequirements = [],
+    statusFields = [], incidentSteps = [], incidentPriorityFactors = [], degradedOperationFields = [],
+    operationsDomains = [], supportModelTypes = [], contextualSupportFields = [], operationalRequestTypes = [],
+    supportKnowledgeRequirements = [], adoptionLifecycleStates = [], adoptionPersonas = [],
+    adoptionValuePropositions = [], onboardingFields = [], migrationFields = [], enablementChannels = [],
+    feedbackSources = [], aiAgentOutcomeFields = [], productivityFields = [], platformValueFields = [],
+    adoptionMeasures = [], operationalMeasures = [], roadmapDecisionFactors = [], backlogItemFields = [],
+    governanceAreas = [], maturityLevels = [], antiPatterns = [], architecturalRules = [], futureCapabilities = [],
+    productionInternalProduct = true, explicitOwnership = true, serviceCommitmentsByCriticality = true,
+    outcomeConnectedHealth = true, integratedOperations = true, alternateCriticalPaths = true,
+    contextualSupportKnowledge = true, ownedPavedPathLifecycles = true, protectedRecoveryCapacity = true,
+    governsSuppliersCostEvidenceDebt = true, adoptionAsJourney = true, onboardingMigrationTraining = true,
+    measuresSafeValue = true, measuresAiOutcomes = true, protectsDeveloperPrivacy = true,
+    tenantPropertyImpactVisible = true, feedbackImprovesRoadmap = true, balancesGrowthReliabilityRetirement = true,
+    vendorNeutralTechnologyIndependent = true, replacesArch011 = false, treatsPlatformAsToolCollection = false,
+    measuresOnlyToolUptime = false, assumesAdoptionFromAccess = false, mandatesUnsupportedPaths = false,
+    makesPlatformOwnEveryAppIncident = false, hidesDegradation = false, singlePortalForCriticalRecovery = false,
+    removesControlsForCost = false, measuresProductivityByCodeVolume = false,
+    measuresAiSuccessByAutonomyVolume = false, usesTelemetryForSurveillance = false,
+    deprecatesWithoutMigration = false, hidesTenantPropertyImpact = false
+  }) {
+    this.platformName = platformName;
+    for (const [key, value] of Object.entries({ objectives, principles, productOwnershipFields, ownershipRoles,
+      serviceCommitmentFields, criticalCapabilities, commandControlFields, controlLoopStages, healthFields,
+      healthEndpointTypes, observabilityCorrelationFields, signalTypes, logRequirements, statusFields, incidentSteps,
+      incidentPriorityFactors, degradedOperationFields, operationsDomains, supportModelTypes, contextualSupportFields,
+      operationalRequestTypes, supportKnowledgeRequirements, adoptionLifecycleStates, adoptionPersonas,
+      adoptionValuePropositions, onboardingFields, migrationFields, enablementChannels, feedbackSources,
+      aiAgentOutcomeFields, productivityFields, platformValueFields, adoptionMeasures, operationalMeasures,
+      roadmapDecisionFactors, backlogItemFields, governanceAreas, maturityLevels, antiPatterns, architecturalRules,
+      futureCapabilities })) {
+      this[key] = Object.freeze([...value]);
+    }
+    for (const [key, value] of Object.entries({ productionInternalProduct, explicitOwnership,
+      serviceCommitmentsByCriticality, outcomeConnectedHealth, integratedOperations, alternateCriticalPaths,
+      contextualSupportKnowledge, ownedPavedPathLifecycles, protectedRecoveryCapacity,
+      governsSuppliersCostEvidenceDebt, adoptionAsJourney, onboardingMigrationTraining, measuresSafeValue,
+      measuresAiOutcomes, protectsDeveloperPrivacy, tenantPropertyImpactVisible, feedbackImprovesRoadmap,
+      balancesGrowthReliabilityRetirement, vendorNeutralTechnologyIndependent, replacesArch011,
+      treatsPlatformAsToolCollection, measuresOnlyToolUptime, assumesAdoptionFromAccess, mandatesUnsupportedPaths,
+      makesPlatformOwnEveryAppIncident, hidesDegradation, singlePortalForCriticalRecovery, removesControlsForCost,
+      measuresProductivityByCodeVolume, measuresAiSuccessByAutonomyVolume, usesTelemetryForSurveillance,
+      deprecatesWithoutMigration, hidesTenantPropertyImpact })) {
+      this[key] = Boolean(value);
+    }
+    Object.freeze(this);
+  }
+}